extract functions out of slice into utils
